Move hint state to Display
