+++ alacritty.yml
@@ -474,7 +474,7 @@
 # and pipe it to other applications.
 hints:
   # Keys used for the hint labels
-  alphabet: "jfkdls;ahgurieowpq"
+  #alphabet: "jfkdls;ahgurieowpq"
 
   # List with all available hints
   #
@@ -487,7 +487,7 @@ hints:
   #    binding:
   #      key: G
   #      mods: Control|Shift
-  enabled: []
+  #enabled: []
 
 # Mouse bindings
 #
