+++ alacritty/src/input.rs
@@ -31,8 +31,9 @@ use alacritty_terminal::vi_mode::ViMotion;
 use crate::clipboard::Clipboard;
 use crate::config::{Action, BindingMode, Config, Key, SearchAction, ViAction};
 use crate::daemon::start_daemon;
+use crate::display::HintState;
 use crate::display::window::Window;
-use crate::event::{ClickState, Event, Mouse, TYPING_SEARCH_DELAY, HintState};
+use crate::event::{ClickState, Event, Mouse, TYPING_SEARCH_DELAY};
 use crate::message_bar::{self, Message};
 use crate::scheduler::{Scheduler, TimerId};
 use crate::url::{Url, Urls};
